# recipe-management-openapi.yaml
...
openapi: 3.1.0
info:
  title: Recipe Management Service API
  description: |
    Comprehensive recipe management service for the Recipe Web Application ecosystem.

    Provides complete functionality for managing recipes, ingredients, steps, tags,
    reviews, and media content. Built with Spring Boot 3.5.3 and Java 21.

    **Features:**
    - Complete recipe CRUD operations
    - Advanced search and filtering
    - Ingredient management with scaling and shopping lists
    - Step-by-step cooking instructions
    - User reviews and ratings
    - Tag-based categorization
    - Media upload and management
    - Revision history tracking
    - JWT authentication integration
  version: 1.0.0
  contact:
    name: Recipe Web App Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1/recipe-management
    description: Local development server
  - url: http://sous-chef-proxy.local/api/v1/recipe-management
    description: Kubernetes deployment
security:
  - BearerAuth: []
...
paths:
  /recipes:
    get:
      tags: [recipes]
      summary: Get all recipes
      description: |
        Retrieves all recipes with pagination support. Returns basic recipe information
        including title, description, ratings, and metadata.
      operationId: getAllRecipes
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: List of recipes retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchRecipesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [recipes]
      summary: Create new recipe
      description: |
        Creates a new recipe with ingredients, steps, and tags. The recipe is
        associated with the authenticated user.
      operationId: createRecipe
      requestBody:
        description: Recipe creation data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeRequest"
      responses:
        "201":
          description: Recipe created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token authentication. Obtain token from user-management-service
        and include in Authorization header as 'Bearer <token>'.
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (0-based)
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      description: Page size
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: Sort criteria (property,direction)
      required: false
      schema:
        type: array
        items:
          type: string
        example: ["createdAt,desc", "title,asc"]
...
  schemas:
    RecipeDto:
      type: object
      required:
        - recipeId
        - title
        - servings
        - createdAt
        - userId
      properties:
        recipeId:
          type: integer
          format: int64
          description: Unique recipe identifier
          example: 123
        userId:
          type: string
          format: uuid
          description: UUID of the user who created the recipe
          example: "123e4567-e89b-12d3-a456-426614174000"
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Recipe title
          example: "Chocolate Chip Cookies"
        description:
          type: string
          maxLength: 2000
          description: Recipe description
          example: "Classic homemade chocolate chip cookies"
        originUrl:
          type: string
          description: The origin URL of the recipe
          example: "https://example.com/recipe/123"
        servings:
          type: number
          format: decimal
          minimum: 0
          description: Number of servings as decimal
          example: 4.5
        preparationTime:
          type: integer
          minimum: 0
          description: Preparation time in minutes
          example: 15
        cookingTime:
          type: integer
          minimum: 0
          description: Cooking time in minutes
          example: 12
        difficulty:
          $ref: "#/components/schemas/DifficultyLevel"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2024-08-28T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-08-28T10:30:00Z"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredientDto"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/RecipeStepDto"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/RecipeTagDto"
        revisions:
          type: array
          items:
            $ref: "#/components/schemas/RecipeRevisionDto"
        favorites:
          type: array
          items:
            $ref: "#/components/schemas/RecipeFavoriteDto"
        comments:
          type: array
          description: Comments on the recipe
          items:
            $ref: "#/components/schemas/RecipeCommentDto"
    RecipeIngredientDto:
      type: object
      required:
        - ingredientId
        - ingredientName
        - quantity
        - unit
      properties:
        recipeId:
          type: integer
          format: int64
          description: Recipe identifier
          example: 123
        ingredientId:
          type: integer
          format: int64
          description: Ingredient identifier
          example: 789
        ingredientName:
          type: string
          minLength: 1
          maxLength: 100
          description: Ingredient name
          example: "All-purpose flour"
        quantity:
          type: number
          format: decimal
          exclusiveMinimum: 0
          description: Ingredient quantity as decimal
          example: 2.25
        unit:
          $ref: "#/components/schemas/IngredientUnit"
        isOptional:
          type: boolean
          description: Whether the ingredient is optional
          example: false
        comments:
          type: array
          description: Comments associated with this ingredient
          items:
            $ref: "#/components/schemas/IngredientCommentDto"
    RecipeStepDto:
      type: object
      required:
        - stepId
        - stepNumber
        - instruction
      properties:
        stepId:
          type: integer
          format: int64
          description: Step identifier
          example: 101
        recipeId:
          type: integer
          format: int64
          description: Recipe identifier
          example: 123
        stepNumber:
          type: integer
          minimum: 1
          description: Step sequence number
          example: 1
        instruction:
          type: string
          minLength: 1
          maxLength: 1000
          description: Step instructions
          example: "Preheat oven to 375°F (190°C)"
        optional:
          type: boolean
          description: Whether this step is optional
          example: false
        timerSeconds:
          type: integer
          minimum: 0
          description: Timer in seconds for this step
          example: 300
        createdAt:
          type: string
          format: date-time
          description: The creation timestamp
          example: "2024-08-28T10:30:00Z"
    RecipeTagDto:
      type: object
      required:
        - tagId
        - name
      properties:
        tagId:
          type: integer
          format: int64
          description: The tag ID
          example: 201
        name:
          type: string
          minLength: 1
          maxLength: 50
          description: The tag name
          example: "dessert"
...
    IngredientCommentDto:
      type: object
      required:
        - commentId
        - recipeId
        - userId
        - commentText
        - createdAt
      properties:
        commentId:
          type: integer
          format: int64
          description: Comment identifier
          example: 301
        recipeId:
          type: integer
          format: int64
          description: Recipe identifier
          example: 123
        userId:
          type: string
          format: uuid
          description: UUID of user who wrote the comment
          example: "123e4567-e89b-12d3-a456-426614174000"
        commentText:
          type: string
          minLength: 1
          maxLength: 1000
          description: Comment text content
          example: "Make sure to sift for best results"
        isPublic:
          type: boolean
          description: Whether comment is public or private
          example: true
        createdAt:
          type: string
          format: date-time
          description: Comment creation timestamp
          example: "2024-08-28T10:35:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-08-28T10:35:00Z"
    RecipeFavoriteDto:
      type: object
      required:
        - recipeId
        - userId
        - favoritedAt
      properties:
        recipeId:
          type: integer
          format: int64
          description: Recipe identifier
          example: 123
        userId:
          type: string
          format: uuid
          description: UUID of user who favorited the recipe
          example: "123e4567-e89b-12d3-a456-426614174000"
        favoritedAt:
          type: string
          format: date-time
          description: Timestamp when recipe was favorited
          example: "2024-08-28T10:35:00Z"
...
    RecipeRevisionDto:
      type: object
      required:
        - revisionId
        - recipeId
        - userId
        - revisionCategory
        - revisionType
        - createdAt
      properties:
        revisionId:
          type: integer
          format: int64
          description: Revision identifier
          example: 501
        recipeId:
          type: integer
          format: int64
          description: Recipe identifier
          example: 123
        userId:
          type: string
          format: uuid
          description: UUID of user who made the revision
          example: "123e4567-e89b-12d3-a456-426614174000"
        revisionCategory:
          $ref: "#/components/schemas/RevisionCategory"
        revisionType:
          $ref: "#/components/schemas/RevisionType"
        previousData:
          type: string
          description: Previous data before revision
          example: "Original recipe data"
        newData:
          type: string
          description: New data after revision
          example: "Updated recipe data"
        changeComment:
          type: string
          description: Comment describing the change
          example: "Updated serving size"
        createdAt:
          type: string
          format: date-time
          description: Revision creation timestamp
          example: "2024-08-28T10:35:00Z"
...
    RecipeCommentDto:
      type: object
      required:
        - commentId
        - recipeId
        - userId
        - commentText
        - isPublic
        - createdAt
      properties:
        commentId:
          type: integer
          format: int64
          description: Unique comment identifier
          example: 501
        recipeId:
          type: integer
          format: int64
          description: Recipe identifier
          example: 123
        userId:
          type: string
          format: uuid
          description: UUID of the user who wrote the comment
          example: "123e4567-e89b-12d3-a456-426614174000"
        commentText:
          type: string
          minLength: 1
          maxLength: 2000
          description: Comment text content
          example: >-
            This recipe is absolutely delicious! I made it for my family
            and everyone loved it.
        isPublic:
          type: boolean
          description: Whether the comment is publicly visible
          example: true
        createdAt:
          type: string
          format: date-time
          description: Comment creation timestamp
          example: "2024-08-28T10:40:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-08-28T10:40:00Z"
    CreateRecipeRequest:
      type: object
      required:
        - title
        - description
        - servings
        - ingredients
        - steps
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Recipe title
          example: "Chocolate Chip Cookies"
        description:
          type: string
          maxLength: 2000
          description: Recipe description
          example: "Classic homemade chocolate chip cookies"
        servings:
          type: number
          format: decimal
          minimum: 1
          maximum: 1000
          description: Number of servings
          example: 4
        preparationTime:
          type: integer
          minimum: 0
          description: Preparation time in minutes
          example: 15
        cookingTime:
          type: integer
          minimum: 0
          description: Cooking time in minutes
          example: 12
        difficulty:
          $ref: "#/components/schemas/DifficultyLevel"
        ingredients:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateRecipeIngredientRequest"
        steps:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateRecipeStepRequest"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/RecipeTagDto"
    CreateRecipeIngredientRequest:
      type: object
      required:
        - ingredientName
        - quantity
        - unit
      properties:
        ingredientId:
          type: integer
          format: int64
          description: The ingredient ID (optional, for referencing existing ingredients)
          example: 789
        ingredientName:
          type: string
          minLength: 1
          maxLength: 100
          description: The ingredient name
          example: "All-purpose flour"
        quantity:
          type: number
          format: decimal
          exclusiveMinimum: 0
          description: The quantity of the ingredient
          example: 2.25
        unit:
          $ref: "#/components/schemas/IngredientUnit"
        isOptional:
          type: boolean
          description: Whether the ingredient is optional
          example: false
        notes:
          type: string
          maxLength: 500
          description: Notes about the ingredient
          example: "Sifted"
    CreateRecipeStepRequest:
      type: object
      required:
        - stepNumber
        - instruction
      properties:
        stepNumber:
          type: integer
          minimum: 1
          description: Step sequence number
          example: 1
        instruction:
          type: string
          minLength: 1
          maxLength: 1000
          description: Step instructions
          example: "Preheat oven to 375°F (190°C)"
        optional:
          type: boolean
          description: Whether this step is optional
          example: false
        timerSeconds:
          type: integer
          minimum: 0
          description: Timer in seconds for this step
          example: 300
...
    SearchRecipesResponse:
      type: object
      required:
        - content
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/RecipeDto"
        totalElements:
          type: integer
          format: int64
          description: Total number of matching recipes
        totalPages:
          type: integer
          description: Total number of pages
        first:
          type: boolean
          description: Whether this is the first page
        last:
          type: boolean
          description: Whether this is the last page
        numberOfElements:
          type: integer
          description: Number of elements in current page
...
    DifficultyLevel:
      type: string
      enum: [BEGINNER, EASY, MEDIUM, HARD, EXPERT]
      description: Recipe difficulty level
      example: EASY
...
    IngredientUnit:
      type: string
      enum:
        - G
        - KG
        - OZ
        - LB
        - ML
        - L
        - CUP
        - TBSP
        - TSP
        - PIECE
        - CLOVE
        - SLICE
        - PINCH
        - CAN
        - BOTTLE
        - PACKET
        - UNIT
      description: Ingredient measurement unit
      example: CUP
...
    RevisionType:
      type: string
      enum: [ADD, UPDATE, DELETE]
      description: Type of revision change
      example: UPDATE
    RevisionCategory:
      type: string
      enum: [INGREDIENT, STEP]
      description: Category of revised entity
      example: INGREDIENT
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
        - path
        - status
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp of when the error occurred
          example: "2024-08-28T10:30:00.000"
        status:
          type: integer
          description: HTTP status code
          example: 404
        error:
          type: string
          description: Error title
          example: "Not Found"
        message:
          type: string
          description: Error message
          example: "Recipe with ID 123 not found"
        path:
          type: string
          description: Request path that caused the error
          example: "/recipes/123"
        requestId:
          type: string
          description: Request ID for tracking
          example: "550e8400-e29b-41d4-a716-446655440000"
        details:
          type: object
          additionalProperties:
            type: string
          description: Optional validation details
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 400
            error: "Bad Request"
            message: "Invalid request parameters"
            timestamp: "2024-08-28T10:30:00Z"
            path: "/recipes"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 401
            error: "Unauthorized"
            message: "JWT token is missing or invalid"
            timestamp: "2024-08-28T10:30:00Z"
            path: "/recipes"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 422
            error: "Validation Error"
            message: "Request validation failed"
            timestamp: "2024-08-28T10:30:00Z"
            path: "/recipes"
            details:
              title: "Title is required"
              servings: "Servings must be positive"
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            status: 500
            error: "Internal Server Error"
            message: "An unexpected error occurred"
            timestamp: "2024-08-28T10:30:00Z"
            path: "/recipes"
...